+++ fj-doc-mod-poi/src/test/java/test/org/fugerit/java/doc/mod/poi/coverage/TestPoiCoverage.java
@@ -41,6 +41,7 @@ public class TestPoiCoverage {
 		Arrays.asList( TEST_LIST ).stream().forEach( c -> {
 			Assert.assertTrue( this.worker( "coverage/xml/"+c+".xml" ) );
 		} );
+		Assert.assertTrue( Boolean.TRUE );
 	}
 	
 }