+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.7.2] - 2024-08-28
+
 ### Changed
 
 - [fj-doc-maven-plugin] added init fugerit venus goal